requirements.txt
# Основные зависимости
paho-mqtt>=1.6.1
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
flask>=3.0.0
protobuf>=4.24.0

# Дополнительные зависимости
scapy>=2.5.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Для разработки
jupyter>=1.0.0
pytest>=7.4.0
